src/third_party/blink/web_tests/fast/workers/worker-performance-time.html

<!DOCTYPE html>
<title>Test the timing in the worker to be consistent between animation and performance.now().</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script id="worker" type="text/worker">
var countWorker = 10;
function animate(time) {
  var timeNow = performance.now();
  countWorker--;
  self.postMessage([time,timeNow, countWorker]);
  if(countWorker > 0)
    requestAnimationFrame(animate);
};

self.onmessage = function(msg){
  requestAnimationFrame(animate);
}
</script>
<script>
async_test(function(t) {
  var blob = new Blob([document.getElementById("worker").textContent]);
  var worker = new Worker(URL.createObjectURL(blob));
  worker.onmessage = t.step_func(function(pairTime) {
    var time = pairTime.data[0];
    var timeNow = pairTime.data[1];
    var count = pairTime.data[2];
    assert_approx_equals(time, timeNow, 1000, "Times must be close enough");
    if(count == 0)
      t.done();
  });
  worker.postMessage("");
}, 'Test the timing in the worker to be consistent between animation and performance.now');
</script>
